+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 - Fix a bug where using `ReadOnly` attribute on hidden field still applying its height [[#51](https://github.com/DarkRewar/BaseTool/issues/51)]
 - Fix a bug that blocks inputs on the dev console on Unity 6 [[#50](https://github.com/DarkRewar/BaseTool/issues/50)]
+- Update the `Console` to work with a settings file [[#57](https://github.com/DarkRewar/BaseTool/issues/57)]
 
 ## 0.3.0